// OpenData-AQI-Frontend/src/app/model/Enum/EMap.ts
export enum enum_MarkerColor {
  well = 'green',
  well_fillColor = '#2a6',
  general = 'yellow',
  general_fillColor = '#fe3',
  warning = 'orange',
  warning_fillColor = '#fa3',
  unhealthy = 'red',
  unhealthy_fillColor = '#d13',
  veryUnhealthy = 'purple',
  veryUnhealthy_fillColor = '#709',
  harm = 'black',
  harm_fillColor = '#000',
}

export enum enum_CardBackgroundColor {
    well = 'rgba(110, 219, 110, 0.356)',
    general = 'rgba(255, 248, 145, 0.692)',
    warning = 'rgba(255, 189, 145, 0.692)',
    unhealthy = 'rgba(255, 145, 145, 0.692)',
    veryUnhealthy = 'rgba(102, 58, 151, 0.623)',
    harm = 'rgba(0, 0, 0, 0.459)'
  }